# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Долгое наступление — Book Club API",
    "description": "API сайта книжного клуба «Долгое наступление». Runtime-каталог хранится в Postgres. Авторизация через NextAuth v5 с session cookie; /api/admin/* требуют session.user.isAdmin=true; /api/cron/* требуют Bearer CRON_SECRET; /api/test/* доступны только в NEXTAUTH_TEST_MODE=true вне production.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://www.slowreading.club",
      "description": "Production"
    },
    {
      "url": "https://book-club-slow-rising.vercel.app",
      "description": "Vercel fallback"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "NextAuth session cookie"
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET для cron endpoints"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Signup": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "contacts": {
            "type": "string",
            "nullable": true
          },
          "selectedBooks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "selectedBookIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prioritiesSet": {
            "type": "boolean"
          }
        }
      },
      "BookPriority": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string"
          },
          "bookName": {
            "type": "string"
          },
          "rank": {
            "type": "integer"
          }
        },
        "required": [
          "bookId",
          "bookName",
          "rank"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Записаться на выбранные книги",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "contacts",
                  "selectedBookIds"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "contacts": {
                    "type": "string"
                  },
                  "selectedBookIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/priorities": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Приоритеты книг пользователя",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Priorities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookPriority"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Сохранить порядок приоритетов",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bookIds"
                ],
                "properties": {
                  "bookIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin dashboard data",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Signups grouped by users and books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Signup"
                      }
                    },
                    "byBook": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Signup"
                        }
                      }
                    }
                  },
                  "required": [
                    "users",
                    "byBook"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/remove-book": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Удалить книгу из записи пользователя",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "bookId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "bookId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/signup-books/{bookId}/status": {
      "patch": {
        "tags": [
          "Signups"
        ],
        "summary": "Обновить personal_status книги в списке пользователя",
        "description": "Позволяет пользователю отметить книгу как 'читаю сейчас' или 'прочитал(а)'. Книги с выставленным статусом исключаются из матчинга как кандидаты для новой группы. Значение null возвращает книгу в активный список.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "nullable": true,
                    "enum": ["reading", "read", null],
                    "description": "null = В списке (активный кандидат), 'reading' = Читаю сейчас, 'read' = Прочитал(а)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный статус",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Пользователь не записан на эту книгу"
          }
        }
      }
    },
...
  }
}